Cleanup old video files

+++ scripts/linux/delete-existingvideo.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
 # Clean any existing videos in case a VM is reused for a build
-rm "$MEDIA_PATH/*.avi"
-rm "$MEDIA_PATH/*.mp4"
+if [ -d "$MEDIA_PATH" ]; then
+  rm "$MEDIA_PATH/*.avi"
+  rm "$MEDIA_PATH/*.mp4"
+fi
